binary lifting bugfix

+++ binary_lifting(doubling)/doubling.hpp
@@ -13,7 +13,7 @@ struct BinaryLifting
     BinaryLifting() : N(0), lgD(0) {}
     BinaryLifting(const std::vector<int> &vec_nxt, int INVALID = -1, int lgd = 0) : N(vec_nxt.size()), INVALID(INVALID), lgD(lgd)
     {
-        while ((1 << lgD) < N) lgD++;
+        while ((1LL << lgD) < N) lgD++;
         mat.assign(lgD, std::vector<int>(N, INVALID));
         mat[0] = vec_nxt;
         for (int i = 0; i < N; i++) if (mat[0][i] < 0 or mat[0][i] >= N) mat[0][i] = INVALID;
@@ -21,9 +21,9 @@ struct BinaryLifting
             for (int i = 0; i < N; i++) if (mat[d][i] != INVALID) mat[d + 1][i] = mat[d][mat[d][i]];
         }
     }
-    int kth_next(int now, int k)
+    int kth_next(int now, long long k)
     {
-        if (k >= (1 << lgD)) exit(8);
+        if (k >= (1LL << lgD)) exit(8);
         for (int d = 0; k and now != INVALID; d++, k >>= 1) if (k & 1) now = mat[d][now];
         return now;
     }
